+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `FormViewValidatorDelegate` implemented.
 - `DateExtension` and `DateFormatterExtension` methos implemented.
 - `DefaultOKChainResponse` implemented.
+- `ActionButton` implemented.
 
 # 0.11.0